chore: append commit to tx log
